+++ scitos2_modules/CHANGELOG.rst
@@ -2,6 +2,14 @@
 Changelog for package scitos2_modules
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2.1.0 (27-03-2025)
+------------------
+* Replaced cmd_vel topic from Twist to TwistStamped in Drive module.
+
+2.0.1 (13-03-2025)
+------------------
+* Added imu_topic parameter to IMU module.
+
 2.0.0 (06-02-2025)
 ------------------
 * Import modules from scitos_mira package.